backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Azure Cosmos DB
azure-cosmos==4.5.1

# Data validation and serialization
pydantic>=2.8.0
pydantic-settings>=2.0.0

# CORS middleware (included in FastAPI but explicitly listed)
# HTTP client for potential external API calls
httpx==0.25.2

# Logging and monitoring
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Type hints support for older Python versions
typing-extensions>=4.12.2

# Date and time handling
python-dateutil==2.8.2

# For testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing HTTP endpoints

# OpenAI
openai>=1.42.0
